+++ src/styles/components/Timer.styled.tsx
@@ -2,7 +2,6 @@ import styled from 'styled-components';
 
 const StyledTimer = styled.div`
   display: inline-block;
-  outline: 1px solid #EEE;
 `;
 
 export default StyledTimer;